Move logic into app.rs
